+++ src/main/java/com/trepicoder/adventurecraft/utils/IBlockPosCheck.java
@@ -1,0 +1,8 @@
+package com.trepicoder.adventurecraft.utils;
+
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IWorld;
+
+public interface IBlockPosCheck {
+    public boolean matchesBlock(IWorld worldIn, BlockPos pos);
+}